Remove all of Redis corporate branded references (#17)
I did a pass to attempt to remove reference to Redis corporate
references in the documentation pages. This includes stack, the stack
modules, insights, and enterprise features. The ones that are still
left:
1. NRedisStack, I don't know enough about StackExchange to know if it's
just a drop in replacement to StackExchange. Maybe something for someone
else to follow up on.
2. References in the modules directory, I'm okay leaving those there for
now.

---------

Signed-off-by: Madelyn Olson <madelyneolson@gmail.com>