Restored topics structure (#61)
This massive change restores the `topics/page.md` structure of the doc
repo, as it was a few years ago.

This is done using a script that is shared in the PR in a comment. This is
what it does:

1. The "aliases" metadata field in each markdown file is used to find
the original "topics" filename. This is attempted for all files in the
repo with a metadata section.
2. For files that don't have a "topics/" alias, pick a topics name for
it, like this:
   * Do this only for files under "docs/".
* If the file path matches "..../NAME/_index.md", the new name is
"topics/NAME.md"
* Otherwise, the file path matches "..../NAME" and then the new name is
"topics/NAME". (This case covers png files and other non-markdown files
too.)
* If the filename includes 'redis', this is replaced by 'valkey'.
Examples:
`docs/getting-started/installation/install-redis-from-source.md` ->
`topics/install-valkey-from-source.md`
* (Special case: If the target file name ends up "topics.md", use
"topics/index.md" instead.)
3. Rename the files to the new names. Commit the the renames.
4. Update all links to files (renamed and not renamed), using all
aliases discovered using all files with a metadata section.
* The links are changed to relative links including the `.md` suffix.
This makes it possible to click around in the github web interface. It's
also good in other situations. For example, users can deploy the docs
(simply .md -> .html) and don't need to deploy the docs in the root
directory.
   * Examples:
* Change link: /docs/reference/cluster-spec => cluster-spec.md (in
topics/protocol.md)
* Change link: /topics/acl => ../topics/acl.md (in commands/auth.md)

Additional changes:

* Fix broken link in cluster tutorial
* Remove some stub files with no real content (commands/_index.md, resources/_index.md)

Fixes #30

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>