Add information about 'server' domain object for LUA (#85)
Documents the behavior for the new server top level domain object and
the new Lua fields. Also updates all of the examples to use the server
object. Also includes some cleanup to references to "Valkey
serialization protocol" which doesn't exist.

There will be some minimal conflicts with
https://github.com/valkey-io/valkey-doc/pull/22, as it attempts to
remove the versioning references which this file also does for the files
that were touched.

---------

Signed-off-by: Madelyn Olson <madelyneolson@gmail.com>
Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>
Co-authored-by: Viktor Söderqvist <viktor.soderqvist@est.tech>