Version 8 man pages preparations (#176)
Move valkey.conf man page to section 5 (was 4)

Add DATE and VERSION variables in Makefile. These
are used in the man page footers.

If running in a git repo, the date and version is taken
from the git repo and if an exact git tag is checked out,
there's a check added that checks that the declared version
matches the git tag. This is to help finding mistake in the
release process. (Run make after git tag, before pushing
the new tag.)

Generate and install gzipped man pages.

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>