Corrections to migration guide (#184)
- User the term "Redis OSS" and "open source Redis", rather than "Redis
Community Edition". The latter refers to the source-available versions
from 7.4 and later.
- Mention that Valkey is backward compatible with all old Redis
versions.
- Update the migration table to include Redis OSS 2.x - 7.2.x. Users
have been confused by the table that seems to indicate that it's not
possible to upgrade from Redis 6.2 to Valkey 8.0, but it is.
- Mention that it may be possible to migrate from proprietary Redis, but
that it's not covered by this document. (For example, writing the data
to disk using AOF without preamble and loading in Valkey can probably
work.)

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>
Co-authored-by: Madelyn Olson <madelyneolson@gmail.com>