Creating a getting started guide for bloom (#310)
~~Adding a getting started guide for bloom filters~~

~~This will include the following:~~
~~- Manual Building / installation (Valkey Server start up + Module
Load)~~
~~- Valkey Extension Docker based startup~~
~~- Examples of executing bloom commands~~
~~- Snippet of executing bloom commands using a python script with
valkey-py~~
~~- Snippet / examples of using bloom configurations~~

~~We will need to decide where we will show getting started guides as
currently they are quite hidden in documentation. But for now I have put
under tutorials and FAQ. If we want a more explicit/obvious place let me
know and i will look at making that change~~

After discussion we have decided to move the getting started information
from a separate page.
The sections we will now include is a tutorial, and a better explanation
of the configs for bloom filter. We will save the extension and python
example for separate pages that are yet to be created. When they are
created I can add the sections from before to them.

---------

Signed-off-by: zackcam <zackcam@amazon.com>
Co-authored-by: KarthikSubbarao <karthikrs2021@gmail.com>