Fix link in CLUSTER SYNCSLOTS  (#377)
Fixes a link that's using incorrect markdown, so it's wasn't rendered as
a link.

Additional fix for the man page scripts: links from a command man page
to another are incorrectly rendered as a reference to a topics page
rather than as a command page. The added flag `--is-command-page` for
command pages fixes this.

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>