Improve the doc of replication backlog fields (#383)
The `repl_backlog_size` in `INFO REPLICATION` is actually the value
of the `repl-backlog-size` configuration item.

The `replication.backlog` in `MEMORY STATS` is the actual memory usage
of the replication backlog buffer.

In the past, they were the same because we would immediately allocate
the backlog. But after we introduce the global shared replication
buffer, now they have expiressed different meanings.

Signed-off-by: Binbin <binloveplay1314@qq.com>